Fix for getting the correct initial test case in test pool
The initial test case extracted from the configuration file is not
always correct (such as when there is direct invocation to
crete_make_concolic() within the binary under test). This fix uses the
first parent test case passed from svm-node as the initial test case
which is guaranteed to be correct.